svgs for counter to 15.1

+++ cron.sh
@@ -4,19 +4,21 @@ BASEDIR="/home/counter.opensuse.org/svg"
 LOCAL="/home/counter.opensuse.org/output"
 BINARY_LOCATION="https://raw.github.com/openSUSE/artwork/master/Marketing%20Materials/Events/openSUSE%20Conference/2013-oS-Conference/countdown-banner"
 BINARY_FINAL_DATE=$(date -d "2014-11-04" +%s)
-REMOTE_LOCATION="counter.o.o:/srv/www/vhosts/0b965c33b89fac9ea53708de378f93dcda084d34/opensuse.org/counter/htdocs"
+REMOTE_LOCATION="rsync://community.infra.opensuse.org/countdown"
 
 set -e
 
 VERBOSE=
 RENDER=
 REMOTE=
-while getopts 'vREB' v; do
+GIT_PULL=
+while getopts 'vREBG' v; do
     case $v in
         v) VERBOSE=1;;
         R) REMOTE=1;;
         E) RENDER=1;;
         B) BINARY=1;;
+        G) GIT_PULL=1;;
         *) echo "ERROR: unsupported parameter -$v">&2; exit 1;;
     esac
 done
@@ -24,6 +26,14 @@ shift $(( $OPTIND - 1 ))
 
 RFLAGS=""
 [ -n "$VERBOSE" ] && RFLAGS="$RFLAGS -v"
+
+if [ -n "$GIT_PULL" ]; then
+    cd "$BASEDIR"
+    GFLAGS=""
+    [ -n "$VERBOSE" ] && GFLAGS="$GFLAGS --verbose"
+    git pull --ff-only $GFLAGS
+fi
+
 
 mkdir -p "$LOCAL"
 /bin/rm -f "$LOCAL"/*.png
@@ -37,16 +47,31 @@ if [ -n "$RENDER" ]; then
         optipng "$png" &>/dev/null
     done
     popd >/dev/null
-
+    for i in *-label.svg ; do
+        cp $i "$LOCAL"/${i//-label} ;
+    done
+    ./mkhtml.pl "$LOCAL"/
+    
     for f in *.html *.css *.js; do
         [ -e "$f" ] || continue
         cp -a "$f" "$LOCAL/"
     done
-    pushd "$LOCAL/"
+    pushd "$LOCAL/" > /dev/null
+    shopt -s nullglob
+    copied=
         for i in *-label*.png ; do
-            ln -s $i ${i//-label} ;
+            j=${i//-label}
+            if [ ! -e $j ] ; then
+              ln -s $i ${i//-label} ;
+            fi
+            copied=1
         done
-    popd
+    shopt -u nullglob
+    if [ ! -n "$copied" ]; then
+       echo "There seems to be no generated images. Please check the output of \"./render.py $RFLAGS $LOCAL/\""
+       exit 1
+    fi
+    popd > /dev/null
 fi
 
 if [ -n "$BINARY" ]; then
@@ -67,6 +92,9 @@ fi
 
 if [ -n "$REMOTE" ]; then
     RSFLAGS=""
+    if [ -e ~/.rsync_pass ] ; then
+      source ~/.rsync_pass
+    fi
     [ -n "$VERBOSE" ] && RSFLAGS="$RSFLAGS -v"
     rsync -ap --delete-after $RSFLAGS "$LOCAL/" "$REMOTE_LOCATION/"
 fi
